requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
chromadb==0.4.24
openai==1.12.0
loguru==0.7.2
atlassian-python-api==3.41.9
streamlit==1.31.1
python-dotenv==1.0.1
pydantic==2.6.3